discussion/requirements.txt
blinker==1.8.2
boto3==1.34.149
botocore==1.34.149
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
elastic-transport==8.13.1
elasticsearch==8.14.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
load-dotenv==0.1.0
MarkupSafe==2.1.5
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
